# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Officia MENA ERP – full-stack accounting for petroleum & agri sectors
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List invoices
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [draft, issued, paid, overdue, cancelled]
        - in: query
          name: type
          schema:
            type: string
            enum: [sales, purchase]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get invoice
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    patch:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update invoice
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
...
  /accounts:
    get:
      operationId: listAccounts
      tags: [accounting]
      summary: Chart of accounts
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      operationId: createAccount
      tags: [accounting]
      summary: Create account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /journal-entries:
    get:
      operationId: listJournalEntries
      tags: [accounting]
      summary: List journal entries
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JournalEntry"
    post:
      operationId: createJournalEntry
      tags: [accounting]
      summary: Create journal entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JournalEntryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
...
  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List clients and suppliers
      parameters:
        - in: query
          name: type
          schema:
            type: string
            enum: [client, supplier, both]
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /contacts/{id}:
    get:
      operationId: getContact
      tags: [contacts]
      summary: Get contact
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    patch:
      operationId: updateContact
      tags: [contacts]
      summary: Update contact
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    delete:
      operationId: deleteContact
      tags: [contacts]
      summary: Delete contact
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── HR / payroll ──────────────────────────────────────────────
  /employees:
    get:
      operationId: listEmployees
      tags: [hr]
      summary: List employees
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [hr]
      summary: Create employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [hr]
      summary: Get employee
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    patch:
      operationId: updateEmployee
      tags: [hr]
      summary: Update employee
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /payroll:
    get:
      operationId: listPayrollRuns
      tags: [hr]
      summary: List payroll runs
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PayrollRun"
    post:
      operationId: createPayrollRun
      tags: [hr]
      summary: Run monthly payroll
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayrollRunInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayrollRun"
...
components:
  schemas:
...
    InvoiceLineItem:
      type: object
      required: [productId, description, quantity, unitPrice, vatRate]
      properties:
        productId:   { type: ["integer","null"] }
        description: { type: string }
        quantity:    { type: number }
        unit:        { type: string }
        unitPrice:   { type: number }
        vatRate:     { type: number }
        discount:    { type: ["number","null"] }
    Invoice:
      type: object
      required: [id, number, type, status, contactId, contactName, issueDate, dueDate, subtotal, vatAmount, total, currency, createdAt]
      properties:
        id:          { type: integer }
        number:      { type: string }
        type:        { type: string, enum: [sales, purchase] }
        status:      { type: string, enum: [draft, issued, paid, overdue, cancelled] }
        contactId:   { type: integer }
        contactName: { type: string }
        issueDate:   { type: string }
        dueDate:     { type: string }
        subtotal:    { type: number }
        vatAmount:   { type: number }
        total:       { type: number }
        currency:    { type: string }
        notes:       { type: ["string","null"] }
        lineItems:   { type: array, items: { $ref: "#/components/schemas/InvoiceLineItem" } }
        createdAt:   { type: string }
    InvoiceInput:
      type: object
      required: [type, contactId, issueDate, dueDate, currency, lineItems]
      properties:
        type:      { type: string, enum: [sales, purchase] }
        contactId: { type: integer }
        issueDate: { type: string }
        dueDate:   { type: string }
        currency:  { type: string }
        notes:     { type: string }
        lineItems: { type: array, items: { $ref: "#/components/schemas/InvoiceLineItem" } }
    InvoiceUpdate:
      type: object
      properties:
        status:    { type: string, enum: [draft, issued, paid, overdue, cancelled] }
        dueDate:   { type: string }
        notes:     { type: string }
        lineItems: { type: array, items: { $ref: "#/components/schemas/InvoiceLineItem" } }
...
    Account:
      type: object
      required: [id, code, name, type, balance, createdAt]
      properties:
        id:        { type: integer }
        code:      { type: string }
        name:      { type: string }
        nameAr:    { type: ["string","null"] }
        type:      { type: string, enum: [asset, liability, equity, revenue, expense] }
        balance:   { type: number }
        parentId:  { type: ["integer","null"] }
        createdAt: { type: string }
    AccountInput:
      type: object
      required: [code, name, type]
      properties:
        code:     { type: string }
        name:     { type: string }
        nameAr:   { type: string }
        type:     { type: string, enum: [asset, liability, equity, revenue, expense] }
        parentId: { type: integer }
    JournalLine:
      type: object
      required: [accountId, debit, credit]
      properties:
        accountId:   { type: integer }
        accountName: { type: ["string","null"] }
        debit:       { type: number }
        credit:      { type: number }
        description: { type: ["string","null"] }
    JournalEntry:
      type: object
      required: [id, date, description, lines, status, createdAt]
      properties:
        id:          { type: integer }
        date:        { type: string }
        description: { type: string }
        reference:   { type: ["string","null"] }
        lines:       { type: array, items: { $ref: "#/components/schemas/JournalLine" } }
        status:      { type: string, enum: [draft, posted] }
        createdAt:   { type: string }
    JournalEntryInput:
      type: object
      required: [date, description, lines]
      properties:
        date:        { type: string }
        description: { type: string }
        reference:   { type: string }
        lines:       { type: array, items: { $ref: "#/components/schemas/JournalLine" } }
...
    Contact:
      type: object
      required: [id, name, type, country, createdAt]
      properties:
        id:        { type: integer }
        name:      { type: string }
        nameAr:    { type: ["string","null"] }
        type:      { type: string, enum: [client, supplier, both] }
        email:     { type: ["string","null"] }
        phone:     { type: ["string","null"] }
        country:   { type: string }
        city:      { type: ["string","null"] }
        vatNumber: { type: ["string","null"] }
        crNumber:  { type: ["string","null"] }
        balance:   { type: number }
        createdAt: { type: string }
    ContactInput:
      type: object
      required: [name, type, country]
      properties:
        name:      { type: string }
        nameAr:    { type: string }
        type:      { type: string, enum: [client, supplier, both] }
        email:     { type: string }
        phone:     { type: string }
        country:   { type: string }
        city:      { type: string }
        vatNumber: { type: string }
        crNumber:  { type: string }
    ContactUpdate:
      type: object
      properties:
        name:      { type: string }
        nameAr:    { type: string }
        type:      { type: string, enum: [client, supplier, both] }
        email:     { type: string }
        phone:     { type: string }
        city:      { type: string }
        vatNumber: { type: string }

    # ── HR ──────────────────────────────────────────────────────
    Employee:
      type: object
      required: [id, name, nameAr, position, department, salary, currency, nationality, joinDate, status, createdAt]
      properties:
        id:          { type: integer }
        name:        { type: string }
        nameAr:      { type: string }
        position:    { type: string }
        department:  { type: string }
        salary:      { type: number }
        currency:    { type: string }
        nationality: { type: string }
        iqamaNumber: { type: ["string","null"] }
        joinDate:    { type: string }
        status:      { type: string, enum: [active, inactive, on_leave] }
        createdAt:   { type: string }
    EmployeeInput:
      type: object
      required: [name, nameAr, position, department, salary, currency, nationality, joinDate]
      properties:
        name:        { type: string }
        nameAr:      { type: string }
        position:    { type: string }
        department:  { type: string }
        salary:      { type: number }
        currency:    { type: string }
        nationality: { type: string }
        iqamaNumber: { type: string }
        joinDate:    { type: string }
    EmployeeUpdate:
      type: object
      properties:
        position:   { type: string }
        department: { type: string }
        salary:     { type: number }
        status:     { type: string, enum: [active, inactive, on_leave] }
    PayrollRun:
      type: object
      required: [id, month, year, totalGross, totalDeductions, totalNet, employeeCount, status, createdAt]
      properties:
        id:               { type: integer }
        month:            { type: integer }
        year:             { type: integer }
        totalGross:       { type: number }
        totalDeductions:  { type: number }
        totalNet:         { type: number }
        employeeCount:    { type: integer }
        status:           { type: string, enum: [draft, approved, paid] }
        createdAt:        { type: string }
    PayrollRunInput:
      type: object
      required: [month, year]
      properties:
        month: { type: integer }
        year:  { type: integer }
...